+++ MoyoBusinessAdvisory/Migrations/DataContextModelSnapshot.cs
@@ -313,8 +313,8 @@ namespace MoyoBusinessAdvisory.Migrations
                     b.Property<string>("ClientId")
                         .HasColumnType("nvarchar(450)");
 
-                    b.Property<double>("NumberOfItems")
-                        .HasColumnType("float");
+                    b.Property<int>("NumberOfItems")
+                        .HasColumnType("int");
 
                     b.Property<DateTime?>("OrderDate")
                         .HasColumnType("datetime2");
